Adjust Arm features to allow dependencies between features
In particular we have some code like warp affine where part of the code
can make use of SVE but the remaining can use the I8MM extension.

To save us needing to support all possible combinations of features we
instead define restrictions such that features are disabled if required
prior features are not present. In particular:

* DotProd is required for I8MM.
* DotProd and I8MM are required for SVE.

These are not strict requirements of the architecture, but are true for
all known implementations to date. In case such a system does exist we
also update the cpudetect.c code to turn off features in line with the
requirements outlined above.

Change-Id: I7ab4f5216adf8d13d458041be1aecc5d40c7a619