rtc-svc: Fix to skip mv search for spatial ref
This is a follow-up to:
https://aomedia-review.googlesource.com/c/aom/+/172461
which removed the skipping of scaling the references when
svc->force_zero_mode_spatial_ref is enabled. Removing this
can slow down the spatial SVC encoder by ~10%.

This fix adds that feature back in, and accounts for
dyanmic change in SVC and resized frames.
The check to skip the reference scaling is done per
reference inside av1_scale_references(), based on the
skip_mvsearch_last/gf/altref.

Removed the TODOs.

Change-Id: Id7fc91c706071dea7d8400f09195889f26d614ec