Use 16-bit vectors between h/v filters in highbd_warp_affine_neon
In highbd warp affine we currently perform a widening multiply to
32-bits during the horizontal filter and then maintain this through to
the vertical filter where we do a non-widening 32-bit multiply. This
also means that we must first widen the filter values in the vertical
filter step since these are stored in memory as 16-bit values.

It seems that once the horizontal filter is complete we can instead
narrow back to 16-bits and then perform a second widening multiply in
the vertical step, saving the need to separately widen the filter
values.

This approach should also allow us to make use of the SVE 16-bit dot
product instructions in a future commit.

Benchmarking this on a Neoverse N2 machine with Clang 16 and GCC 12
shows a geomean 5.2% reduction in runtime of the warp affine speed
tests.

Change-Id: I9891d919b6a53edc7e5ca8b6dca1ee23f46ef45a