Remove dist_wtd_convolve_2d_horiz_neon transpose for w = 4
Refactor dist_wtd_convolve_2d_horiz_neon to only use a transpose for
blocks with height and width >= 8. Using a single-row-at-a-time
approach for smaller blocks (particularly those using a 4-tap filter)
is faster since the cost of transposing a smaller block is
proportionally larger than for a bigger one where there are more
independent operations.

Change-Id: I942e132c07a75a2834ab78fb2a335467bfbb43d0