Remove av1_dist_wtd_convolve_x_neon transpose if w = 4 || h = 4
Refactor av1_dist_wtd_convolve_x_neon to only use a transpose for
blocks with height and width >= 8. Using a single-row-at-a-time
approach for smaller blocks (particularly those using a 4-tap filter)
is faster since the cost of transposing a smaller block is
proportionally larger than for a bigger one where there are more
independent operations.

Change-Id: I92fffdbd05c1d51e83037dd52b2af06968b4d094