Remove convolve_2d_sr_horiz_neon transpose for w = 4
Refactor convolve_2d_sr_horiz_neon to only use a transpose for blocks
with height and width >= 8. Using a single-row-at-a-time approach for
smaller blocks (particularly those using a 4-tap filter) is faster
since the cost of transposing a smaller block is proportionally larger
than for a bigger one where there are more independent operations.

Change-Id: I934ff3d91ed6a8fc92f3a9d32c32226e16c20f74