Explain skip_txfm
In this cl, we document how mbmi->skip_txfm is determined in the
encoding (reconstruction) stage.

Without the document, it is difficult to track how this variable is
determined in the encoder.

We remove the additional layer that takes mbmi->skip_txfm as a reference
and modifies it later, making the code much easier to understand.

Change-Id: I13f2326cdb9d6575f3d96f07254a3025ec8e103c