rtc: Fix to color artifacts for screen
Move up source_variance computation to
superblock level, and use it to condition
thresholds for setting color_sensitivity,
on both superblock and coding block level.

And make some fixes and adjustments to how the
color_sensitivity is set in nonrd_pickmode.

This significantly reduces color artifacts
for screen content mode.

Change-Id: Ie21dc3212a6572a325f095e881617a4eeedcf607