rtc: Fix to color artifacts for screen
Make some fixes and adjustments to how the
color_sensitivity is set in nonrd_pickmode.

And move up source_variance computation to
superblock level, and use it to condition
thresholds for setting color_sensitivity
at superblock level.

This significantly reduces color artifacts
for screen content mode.

This is a follow-up fix (mainly for screen) to:
https://aomedia-review.googlesource.com/c/aom/+/178881

Change-Id: Ie21dc3212a6572a325f095e881617a4eeedcf607