Use av1_highbd_convolve_x_sr_c for 2x<h> and <w>x2
Use the scalar C implementation of av1_highbd_convolve_x_sr instead of
the Neon one for 2x<h> and <w>x2 blocks. This allows the Neon paths to
be updated to assume blocks are at least 4 wide and 4 high - removing
the need for tests in the inner loops to check if only 2 elements per
row, or only 2 rows out of 4, should be stored.

Change-Id: I33fecbf2a08210f2c187d477aa2a7add2974789b