Add x86 SIMD for highbd frame error
Add SSE2 and AVX2 implementations of av1_calc_highbd_frame_error().
Also add two sets of tests for these new functions:

* AV1HighbdFrameErrorTest: Check that C and SIMD implementations of
  av1_highbd_frame_error() give the same results for all bit depths

* AV1HighbdFrameErrorConsistencyTest: Check that 8-bit and 16-bit code
  paths are consistent with each other, ie. that av1_frame_error_c()
  and av1_highbd_frame_error_c() give the same results on 8-bit content

Benchmarking suggests that the SSE2 and AVX2 implementations of
av1_calc_highbd_frame_error() have a real-time speedup of 1.4x and 1.9x
respectively, relative to C. For comparison, the existing SIMD for the
lowbd path has a speedup of 1.4x for SSE2 and 1.6x for AVX2.

The limiting factor here appears to be memory latency - these functions
perform one unpredictable table lookup per pixel, albeit from a table
which is only 2KB. But the remaining logic is simple enough that it
xeems this is a significant fraction of the runtime even for the C
implementation.

Change-Id: I11dd5e99c1212257206c8b4d1bd434b84e86f4bb