Use 16-bit cospi/sinpi tables in av1_fwd_txfm2d_neon.c
All of the data from these tables fits in 16-bits, however there is
already a body of code that relies on the data being 32-bits wide, so
duplicate the table.

This commit doesn't attempt to change the actual intrinsics to make use
of the narrower data type but the narrower width of the data being
loaded already provides some benefit by itself.

In particular when benchmarking on a Neoverse V1 machine with Clang 16
and GCC 12 we observe a geomean 3% improvement in the
NEON/AV1FwdTxfm2dTest.DISABLED_Speed tests.

Change-Id: Ia8a93919e92ffbf4f002acc8f8ad295b2b458344