Clean up butterfly helpers in av1_fwd_txfm2d_neon.c
A lot of the helper functions are slight variants of each other, so
avoid this duplication by creating a single helper that the others call
in to for each of the data types/sizes of interest.

We can also get rid some of the helpers entirely by swapping
input/output/coefficients to make them equivalent to other existing
helpers or negating the inputs or outputs to the helper elsewhere.

Change-Id: I406c6db4b41127178403898ec9e296477f3818ab