Unswitch ifs and clean up early exits in av1_fwd_txfm2d_neon.c
The current if conditions at the end of each txfm2d function means that
the code at the end refers to either buf0 or buf1 depending on how it
arrived there. To clean these up we can simply duplicate the trailing
code into the if statement body.

This also allows a minor simplification since the round_shift function
only applies in one of the branches.

While we are here, also clean up the {row,col}_txfm null check: prefer
an early exit to avoid needing an extra level of indentation.

Change-Id: I3c7f307b1b00d2594d2cf4161a07d8eb8132d896