Improvements to VBR rate accuracy trade offs.
The VBR rate control mechanism relies on an estimate of
the Q required to generate a target number of bits.

Better metrics are achieved for a clip if the Q is constrained
within a limited range so the encoder tries to estimate
a suitable min / max q range. However, if the initial estimate
is poor the clip may substantially overshoot or undershoot
the target range.

There are two main adjustment mechanisms to deal with
such cases.

The first allowed for some limited adjustment of the Q range
and the heuristics used to estimate Q after each ARF group.

The second allows for adjustments to the min and max Q
each frame but is only activated if the accumulated error
exceeds a given threshold.  If the overshoot and undershoot
percentage is set to 100 on the command line this second
stage was disabled.

This patch seeks to improve (on average) the accuracy of the
initial Q to rate estimate and to allow more adjustment via
the first (non thresholded) mechanism.

The secondary mechanism is left in place but acts more
slowly and is more limited in terms of the maximum amount
by which the min or max Q can be adjusted.

The net effect of these changes is a huge improvement in
the baseline rate accuracy and the rate accuracy / metrics trade off
when the overshoot and undershoot threshold is set to 100 or to
higher values such as 75% or 50%.

Eg. Old code threshold 100% vs 75%

XX

New code threshold 100% vs old code 75%

XX

A much fuller analysis of the rate accuracy behavior and metrics
trade off will be presented in a separate document.

STATS_CHANGED

Change-Id: I4dacd7c62bb55289fe0cd3d46c455dc90f48e9d9