Improvements to VBR rate accuracy trade offs.
The VBR rate control mechanism relies on an estimate of
the Q required to generate a target number of bits.

Better metrics are achieved for a clip if the Q is constrained
within a limited range so the encoder tries to estimate
a suitable min / max q range. However, if the initial estimate
is poor the clip may substantially overshoot or undershoot
the target range.

There are two main adjustment mechanisms to deal with
such cases.

The first allowed for some limited adjustment of the Q range
and the heuristics used to estimate Q after each ARF group.

The second allows for adjustments to the min and max Q
each frame but is only activated if the accumulated error
exceeds a given threshold.  If this overshoot / undershoot
percentage is set to 100% on the command line, this second
stage is disabled.

This patch seeks to improve (on average) the accuracy of the
initial Q to rate estimate and to allow more adjustment via
the first (non thresholded) mechanism.

The secondary mechanism is left in place but acts more
slowly and is more limited in terms of the maximum amount
by which the min or max Q can be adjusted.

The net effect of these changes is a huge improvement in
the baseline rate accuracy and the rate accuracy / metrics trade off
when the overshoot and undershoot threshold is set to 100 or to
higher values such as 75% or 50%.

Results below show the bdrate change (-ve better) for psnr and vmaf
followed by change in average absolute rate error. This last
number is a absolute change in the % error not a % change of % error.
More detailed results and explanation are given in a seperate
document.

Old code threshold 100% vs 75% (rate error reduced, but metrics
much worse)

Low Res opsnr 1.282%  Vmaf 1.949%  (abs rate error -4.088%)
Ugc 360 opsnr 2.308%  Vmaf 2.474%  (abs rate error -4.817%)
Mid Res opsnr 1.362%  Vmaf 1.314%  (abs rate error -4.210%)
Hd Res  opsnr 2.637%  Vmaf 3.984%  (abs rate error -7.369%)

New code threshold 100% vs old code 100%

Low Res opsnr 0.113%  Vmaf -0.324%  (abs rate error -6.888%)
Ugc 360 opsnr 0.084%  Vmaf -0.492%  (abs rate error -4.121%)
Mid Res opsnr -0.832%  Vmaf -0.937%  (abs rate error -2.772%)
Hd Res  opsnr 0.521%  Vmaf -0.732%  (abs rate error -8.547%)

In summary the worst case (threshold 100%) rate accuracy
with the new patch is as good as the 75% threshold case
previously but instead of a metrics hit of circa 1-4% the metrics
are broadly comparable for opsnr and improved for Vmaf.

Change to level monitoring unit tests to be over one second
(30 frames) as this is the key frame interval for the tests. Otherwise
the level assessment is just done on the last 10 frames as if it were
a whole second.

STATS_CHANGED

Change-Id: I4dacd7c62bb55289fe0cd3d46c455dc90f48e9d9