rtc-screen: Remove sad_thresh for sb_me
Remove the thresh_sad used in the logic to set
sb_me = 1 (superblock ME to be tested in nonrd_pickmode).
The feature sb_me is already conditioned on sad relative
to sad for zero motion, and there is further constraint
in pickmode that sb_me is only used for big blocks. So
additional check on sad does not seem needed.

Stats change on speed 10 screen:
avg/ovr/ssim, IC speedup:
-0.098/-0.320/-0.251, 0.020

Change-Id: Ieb9484ac5b419e2af346b744fa6f4511d9cefe59