Refactor memory alloc/dealloc of CDEF search context buffers
This change adds a pointer to the structure 'CdefSearchCtx'
in 'AV1_COMP' to facilitate the deallocation of CDEF search
context buffers in encoder_destory(), which is required to
avoid potential memory leaks in case of an error during CDEF
search. Also used CHECK_MEM_ERROR() to allocate the CDEF
search context buffers.

Bug: aomedia:3276

Change-Id: Id2a343ae7286f27b94d0c24d3b44af93bd5eb156