Pre-shift butterfly weights in av1_fwd_txfm2d_neon.c
The existing implementation of the butterfly helpers shift right by
`cos_bit` bits after perfoming the multiply by the weights, however we
know that the range of `cos_bit` is only 10 through 13 inclusive and we
know that the weights themselves are relatively small.

Shifts by immediates are cheaper on AArch64 so we can simply always
shift by the maximum amount (13) and multiply by a larger value to
effectively shift the multiply result left into the correct place. This
shift left can be done by pre-shifting the butterfly weights by `13 -
cos_bit`.

Running the txfm2d speed tests on a Neoverse V1 machines with Clang 16
and GCC 12 shows a geomean 5.3% reduction in the times reported.

This reverts commit 6e2e560a2e3fcdc5d4a605f9d2de6c5da77e9e84.

Change-Id: I66a0125308494f3af25ee18e8e041f5965ba8fbb