Fix Wiener filter extreme value tests
The aforementioned test cases fill the whole image with the highest
possible value which is not suitable because the filter uses differences
to the average pixel value, in this case always zero. The fix sets the
source image and dgd values to maximise the difference with the average
and to stress the accumulation parts of the implementations.

Change-Id: I79e1db52b1d537d869c5b7eb769c80040cd52da1