Align an array in highbd_warp_error()
The recently added av1_calc_highbd_frame_error_avx2 function requires
its input to be 32-byte aligned, due to the use of the _mm256_load_si256
intrinsic. Unfortunately, when this is called from highbd_warp_error(),
the input array was not properly aligned, causing a segfault. This is
easily solved by aligning the allocation properly.

Fixes a failure in AV1/EndToEndTestLarge.EndtoEndPSNRTest/41

Change-Id: I9cf3f4b40b324af50c4fcca10deaecae72499b18