Avoid variable shifts in av1_fwd_txfm2d_neon.c
Currently the shift amounts are looked up through av1_fwd_txfm_shift_ls
however this is simply a lookup into a constant array with a constant
index, and therefore the resulting shift constants can be inlined and
the lookup completely removed.

With these shift constants inlined we can now use the shift-by-immediate
instructions, and remove several resulting instances of shifts by zero.

Benchmarking this with GCC12 and Clang16 on a Neoverse V1 machine using
the txfm2d speed tests shows a geomean ~5% reduction in the timings
reported.

Change-Id: I60a7957e18cadf7df27b6e7331567d3a7e8eb2f1