Reveal and propagate constants in av1_fwd_txfm2d_neon.c
The existing code makes use of the following helpers and arrays:

*  get_txw_idx(...);
*  get_txh_idx(...);
*  av1_fwd_cos_bit_col[...][...]
*  av1_fwd_cos_bit_row[...][...]

However since we have a separate implementation for all transpose sizes
these are constant and can be removed. This allows the compiler to
completely unroll a few additional loops and also means we can simplify
a few cases of e.g. `AOMMIN(4, height_div8)` which are now trivial.

Performing these simplifications also means that the value of `cos_bit`
is now constant for a given transform.

Running the txfm2d speed tests with Clang 16 and GCC 12 on a Neoverse V1
machine shows a geomean 1% reduction in times reported, mostly due to a
small but noticeable benefit for the smaller problem sizes such as 4x4.

Change-Id: I47f0633dd5c03a0c2af692d3475cb7bba247541e