Rework round_shift_sqrt2 Neon helpers, use immediate shifts
The existing store_rect_16bit_to_32bit functions use a vector of shift
amounts which are always constant, so just use the immediate form
instead.

Also inline these functions into the store functions since that is their
only use and pull out a common helper function to handle the shifting
itself instead, since this can be used for both the 4-wide as well as
the 8-wide code.

The fidentity functions can also make use of these helpers with some
slight tweaks (such as needing `sqrt(2) * 2` in the fidentity8x16_neon
case). Also use the existing shift helpers in the fidentity functions to
match.

Change-Id: Ia63f47c61141119f0a8fe0dcb8a9144f6e70fd08