Use ud_adjust_input_and_stride in highbd_fwd_txfm_neon.c
Adjusting the pointer and stride based on ud_flip ahead of the code that
actually loads the vectors seems to generate marginally better code for
the vector portions.

We already have this functionality in the lowbd code, so move it into a
common header and use it from there instead.

Running the highbd txfm speed tests with Clang 16 and GCC 12 on a
Neoverse N2 machine shows a geomean 0.5% reduction in times reported.

Change-Id: I02a6f2d71a4e53031cbb87b50eed40dbc50f4ee9