aom_vector_var_neon: fix signed int overflow
The square of mean needs to be calculated using an unsigned int to
accommodate the 32-bit range (510 * 16 * 8 = 65280).

Given the behavior of unsigned 2's complement math, the abs() added in
the conversion of horizontal_add_s16x8() isn't strictly necessary (the
value will roll over in a well-defined manner when dealing with negative
values), but it prevents warnings when using -fsanitize=integer.

Change-Id: Id1d92fe22904e888b6285520532555d867621757