Reveal and propagate constants in highbd_fwd_txfm_neon.c
Similar to a related previous change for the non-highbd code. The
existing code makes use of the following helpers and arrays:

*  get_txw_idx(...);
*  get_txh_idx(...);
*  av1_fwd_cos_bit_col[...][...]
*  av1_fwd_cos_bit_row[...][...]

However since we have a separate implementation for all transpose sizes
these are constants and can be removed. This allows the compiler to
completely unroll a few additional loops and also means we can simplify
a few cases where constants to loop bounds are now trivial.

Performing these simplifications also means that the value of `cos_bit`
is now constant for a given transform, although in some cases we have
kept the array lookup to avoid too many arbitrary constants in the
implementations.

Running the txfm2d speed tests with Clang 16 and GCC 12 on a Neoverse N2
machine shows a geomean 4% reduction in times reported.

Change-Id: I0c10ccf966e41d7a45b540dccbdcdf2dee7bd9fc