Adjust memory ordering in highbd_fwd_txfm_neon.c to match lowbd
Currently most of the kernels in the highbd txfm2d code take a col_num
parameter that indicates both the number of times to apply the transform
and also the stride between successive elements. This complicates the
address calculations in the kernels themselves since we now need to
multiply by the strides rather than simply loading consecutive addresses
in memory.

The non-highbd version of the same code does not take a stride parameter
and assumes all values for the transform kernels are adjacent in memory,
so adjust the memory order in the highbd code to match.

Since the memory order is now pretty straightforward we can also
simplify the transpose functions and move them into transpose_neon.h.

Running the highbd txfm2d speed tests with Clang 16 and GCC 12 on a
Neoverse N2 machine, this commit gives a geomean ~0.7% reduction in
times reported.

Change-Id: I7ea8751bdc33390c5ee83f69e15eec6de8a839d2