Correct the error handling of buffer allocations in encodeframe.c
This CL uses the thread-specific 'x->e_mbd.error_info' for error
reporting when allocating the buffer 'sb_org_stats' in encode_rd_sb().
Also, replaced CHECK_MEM_ERROR() by aom_malloc() in
encode_frame_internal() for the allocation of local buffers so that
the buffers are deallocated correctly in case of an error.

Bug: aomedia:3276
Change-Id: I9e4f223e670898d5151a6ccf07e8f4ea2f01c6e8