Call x86_simd_caps() inside #if !CONFIG_SHARED
All the SIMD-specialized functions (arm or x86) are internal functions,
so they are not exported from the libaom shared library. If
CONFIG_SHARED is 1, it is not necessary to call x86_simd_caps() and
append_negative_gtest_filter() either.

This is a follow-up to commit 374fab3d3c0078f490c51d92153d191eb8b843cc.

Change-Id: I722d53e253bde341aafbb6d9f4a4304565a5caec