Detect encoder and image high bit depth mismatch
If the encoder is initialized with the AOM_CODEC_USE_HIGHBITDEPTH flag,
then the input image must have a high bit depth format. Similarly, if
the encoder is not initialized with the AOM_CODEC_USE_HIGHBITDEPTH flag,
then the input image must not have a high bit depth format.

Make AV1's encoder_init() fail if the AOM_CODEC_USE_HIGHBITDEPTH flag is
specified but CONFIG_AV1_HIGHBITDEPTH is 0.

Bug: aomedia:3489
Change-Id: Ie08f9ec132b5cb85f1465090f6bc771e5ba1e782