rtc: Fix for quality layers svc
Add a flag to indicate whether the current spatial
layer has a quality layer as reference, and if so,
force the zeromv testing of this reference.

For screen, in many cases, the lower quality layer has
good enough prediction/quality that makes it worth
testing this mode (i.e., pull up the quality from the lower layer).

Testing on speed 10 screen: shows reduced artfacts and
faster encoding, as the zeromv-lower layer reference often
is good enough and can trigger skipping of other modes.

Change-Id: If6958be088d91329b4dfb693a72601e212fa702e