rtc: Fix for quality layers svc
Add a flag to indicate whether the current spatial
layer has a quality layer as reference, and if so,
force the zeromv testing of this reference.

In current code a nonzero-mv shift was being tested
due to the phase of the downsampling filter, but for
quality layers there is no shift and zeromv should be
tested.

Testing on speed 10 screen shows reduced artfacts and
faster encoding, as the zeromv-lower layer reference often
is good enough and can trigger skipping of other modes.

Change-Id: If6958be088d91329b4dfb693a72601e212fa702e