Remove stride parameter from kernels in highbd_fwd_txfm_neon.c
With the new simplified memory layout these strides are always the same
size as the kernel transform length, so the parameter can be hard-coded
and removed.

Benchmarking with Clang 16 and GCC 12 on a Neoverse N2 machine using the
highbd txfm2d speed tests, this commit shows a geomean 0.3% reduction in
the times reported.

Change-Id: I6f6040dcb56bc2efe88eb1219dec312861023697