Refactor 4x8 and 8x8 kernels in highbd_fwd_txfm_neon.c
Extract a 4x8 kernel out of the existing fadst8x8_neon and idtx8x8_neon
kernels, and call them in a loop over howmany for the new 8x8
implementations.

With this change the 4x8 kernels are only ever called with howmany=1, so
the parameter can be removed.

Additionally remove idtx32x8_neon, since it is now a duplicate of
idtx8x8_neon.

Benchmarking this change using the txfm2d speed tests with Clang 16 and
GCC 12 on a Neoverse N2 machine shows a geomean 0.5% reduction in times
reported, primarily due to the 8x8 cases getting noticeably faster when
compiled with GCC.

Change-Id: If4dca4a0d7a865abe975d01cc54945af38c1a78d