Free mv_costs and dv_costs in encoder_destroy()
This CL frees 'mv_costs' and 'dv_costs' in
free_thread_data() to prevent memory leaks in case of
early termination. In order to avoid double free error in
case of 'x->mv_costs', the allocation and freeing of the
same is done using a newly introduced 'mv_costs_base_ptr'
in ThreadData structure. 'x->mv_costs' is initialized
appropriately.

Bug: aomedia:3276
Change-Id: I0954513c9c0bdb216cd4fbeaf75533552472132f