Update the film_grain test for monochrome
The current test was only failing for realtime mode
when encoder is initialized with monochrome=1 and then
second frame is encoded with monochrome=0.

But this actually fails for both realtime and good quality mode.
Reason it didn't fail for good quality mode was that
lag_in_frames > 0, and this was causing the test to always set
monochrome=1 for the encoded frames.

In this change:
-Set lag_in_frames = 0 to properly set the monochrome flag
 per encoded frame
-Add 4 modes to the test for: monochrome always off, always on,
 and 2 dynamic modes: (0, 1, 0) and (1, 0, 1).
-Use speed 7 for realtime to cover nonrd_pickmode path
-Add test to cover encoding delta frames as well as key frames

Test will fail for 4th mode when monochrome flag is 1, 0, 1,
where encoder is initialized with monochrome=1 and then encoding
on second frame is tried with monochrome=0. This failure should be
expected as encoder should be reset/re-alloc on this flip.

Bug:aomedia:3465
Change-Id: I4fa982e86f172f4ee8d27e5b6b9f8f4f6aa68998