Avoid loop around small kernel calls in highbd_fwd_txfm_neon.c
For the 4x8, 4x16, 8x4, 16x4, and 16x8 highbd txfm2d implementations we
currently loop over the component kernels in some cases and call the
"howmany" implementations in others. This is particularly problematic
since the kernels cannot be inlined due to the txfm kernel array lookup.

We already have a howmany parameter to do multiple iterations of the
kernel in the 16x8 txfm2d case, so use that instead. For the other cases
we introduce new "howmany" kernels that wrap the existing kernels.

To avoid name clashes we also take this opportunity to rename the
component kernel and corresponding array cases. In particular we now
distinguish between cases that process exactly four elements per call
("x4", from the int32x4_t parameter type) versus processing N elements
in a loop ("xn").

Benchmarking this on a Neoverse N2 machine with Clang 16 and GCC 12, the
txfm2d speed tests show a geomean 1.1% reduction in the times reported.
In particular the 16x4 cases improve by 5-15% depending on the txfm type
and compiler.

Change-Id: Ia0ae7c0928c6e29c98b0055dba521d1cd2b6579c