highbd_fwd_txfm_neon.c: Inline load/store helpers into kernels
Currently in cases where the txfm2d component kernels are called via an
array lookup the compiler has no way of inlining these. For the
`load_buffer_*` helper function this is an issue since we load data from
memory and then must immediately store it back to the stack for passing
to the component kernel, and similarly for the `store_buffer_*` calls.

To get around this issue we can inline the `load_buffer_*` and
`store_buffer_*` helpers into the component kernels themselves, creating
specific "col" and "row" kernels for their respective transforms and
avoiding the unnecessary stack spills between calls. This also matches a
similar change previously done for the lowbd Neon code.

Specifically we introduce new `TRANSFORM_COL_{ONE,MANY}` and
`TRANSFORM_ROW{,_RECT}_{ONE,MANY}` macros to define the new wrapper
kernels. The "many" kernels also handle looping over the component
kernels and helpers where we were previously passing a "howmany"
parameter to the component kernel.

To allow using the `store_buffer_*` helpers in the new wrapper kernels,
we also need to adjust the helper to expose a stride parameter rather
than re-using the height from the macro.

Benchmarking this change on a Neoverse N2 machine with Clang 16 and GCC
12, the highbd txfm2d speed tests show a geomean 5.2% reduction in times
reported.

Change-Id: I8c17f9f1bb5f6931593e9f990ec45ce7ca225519